requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
python-dotenv>=1.0
pydantic>=2.7
scikit-learn>=1.4
joblib>=1.3
ollama>=0.2.0
